ml_engine.py
import pandas as pd
import numpy as np
from sklearn.linear_model import LinearRegression
from sklearn.preprocessing import StandardScaler
import warnings
warnings.filterwarnings('ignore')

def predict_future_orders(df, months_ahead=6):
    """
    Predict future order volume using linear regression.
    """
    try:
        # Prepare training data
        monthly = df.groupby(['purchase_year', 'purchase_month']).size().reset_index(name='order_count')
        
        if len(monthly) < 2:
            return {"error": "Insufficient data for prediction"}
        
        # Create numerical features
        monthly['time_index'] = range(len(monthly))
        
        X = monthly[['time_index']].values
        y = monthly['order_count'].values
        
        # Train model
        model = LinearRegression()
        model.fit(X, y)
        
        # Predict
        future_indices = np.array([[len(monthly) + i] for i in range(1, months_ahead + 1)])
        predictions = model.predict(future_indices)
        
        # Format results
        last_year, last_month = monthly.iloc[-1][['purchase_year', 'purchase_month']].astype(int)
        forecast = []
        
        for i, pred in enumerate(predictions):
            month = last_month + i + 1
            year = last_year
            if month > 12:
                month = month % 12
                year += 1
            
            forecast.append({
                "year": int(year),
                "month": int(month),
                "predicted_orders": max(0, int(round(pred))),
                "confidence": 0.85
            })
        
        return {
            "success": True,
            "forecast": forecast,
            "model_accuracy": round(float(model.score(X, y)), 3)
        }
    except Exception as e:
        return {"error": str(e)}

def clustering_analysis(df):
    """
    Analyze clusters in delivery performance.
    """
    try:
        from sklearn.cluster import KMeans
        
        # Prepare features
        features = df[['delivery_days', 'order_purchase_timestamp']].copy()
        features['days_since_start'] = (features['order_purchase_timestamp'] - features['order_purchase_timestamp'].min()).dt.days
        features = features[['delivery_days', 'days_since_start']].dropna()
        
        if len(features) < 3:
            return {"error": "Insufficient data for clustering"}
        
        # Normalize
        scaler = StandardScaler()
        features_scaled = scaler.fit_transform(features)
        
        # Cluster
        kmeans = KMeans(n_clusters=3, random_state=42, n_init=10)
        clusters = kmeans.fit_predict(features_scaled)
        
        # Analyze
        analysis = {
            "success": True,
            "clusters": int(kmeans.n_clusters),
            "inertia": float(kmeans.inertia_),
            "cluster_details": []
        }
        
        for i in range(kmeans.n_clusters):
            cluster_data = features[clusters == i]
            analysis["cluster_details"].append({
                "cluster": i,
                "size": int(len(cluster_data)),
                "avg_delivery_days": round(float(cluster_data['delivery_days'].mean()), 2),
                "std_delivery_days": round(float(cluster_data['delivery_days'].std()), 2)
            })
        
        return analysis
    except Exception as e:
        return {"error": str(e)}

def anomaly_detection(df):
    """
    Detect anomalies in delivery performance.
    """
    try:
        # Calculate delivery metrics
        delivery_days = df['delivery_days'].dropna()
        
        if len(delivery_days) < 5:
            return {"error": "Insufficient data for anomaly detection"}
        
        # Use IQR method
        Q1 = delivery_days.quantile(0.25)
        Q3 = delivery_days.quantile(0.75)
        IQR = Q3 - Q1
        
        lower_bound = Q1 - 1.5 * IQR
        upper_bound = Q3 + 1.5 * IQR
        
        anomalies = df[(df['delivery_days'] < lower_bound) | (df['delivery_days'] > upper_bound)]
        
        return {
            "success": True,
            "total_records": int(len(df)),
            "anomalies_detected": int(len(anomalies)),
            "anomaly_percentage": round((len(anomalies) / len(df)) * 100, 2),
            "lower_bound": round(float(lower_bound), 2),
            "upper_bound": round(float(upper_bound), 2),
            "details": {
                "fast_deliveries": int(len(df[df['delivery_days'] < lower_bound])),
                "slow_deliveries": int(len(df[df['delivery_days'] > upper_bound]))
            }
        }
    except Exception as e:
        return {"error": str(e)}

def correlation_analysis(df):
    """
    Perform correlation analysis on numeric columns.
    """
    try:
        numeric_df = df.select_dtypes(include=[np.number])
        
        if numeric_df.empty:
            return {"error": "No numeric columns found"}
        
        correlation_matrix = numeric_df.corr().to_dict()
        
        return {
            "success": True,
            "correlation_matrix": correlation_matrix,
            "strong_correlations": []
        }
    except Exception as e:
        return {"error": str(e)}

def get_all_ml_insights(df):
    """
    Generate comprehensive ML-based insights.
    """
    return {
        "predictions": predict_future_orders(df, months_ahead=6),
        "clustering": clustering_analysis(df),
        "anomalies": anomaly_detection(df),
        "correlations": correlation_analysis(df)
    }
