requirements.txt
flask>=3.0.0
pandas>=2.2.0
numpy>=2.1.0
scikit-learn>=1.5.0
reportlab>=4.0.0
Pillow>=10.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1